Move serializers to own module
